Mate-stop in one direction only, plus a hygiene pass
The mate-stop counter used |score|, so it fired whether the mate was ours or
the opponent's. The intent stated twelve lines above is one-directional:
"you found the mate, play it instead of thinking further". Receiving mate is
the opposite case — stopping early gains nothing in a lost position and gives
up the only resource left, which is to keep looking for the line that delays
longest or to discover the mate was a pruning artifact. Verified both ways:
giving mate still stops at depth 7, receiving mate now runs to depth 64 where
it previously broke out at iteration 7.
Adds a singular_depth >= 1 guard. Byte-identical with shipped defaults
(mindepth 6, div 2 gives at least 2), so nothing changes today. It exists so
SingularDepthDiv can be tuned without a trap: raise it enough and
singular_depth reaches 0, where td_negamax falls into quiescence — the
singularity check would stop being a search and become a qsearch, changing
meaning rather than degree. A parameter that can invalidate its own test is
not tunable.
Retires NEGEXT_family_laptop.json to _obsolete/. NegExtCut is unreachable
with NegExtOrder=0, and a range of 4 is not SPSA material in any case: theta
is continuous but the engine plays the integer, so it must cross half a unit
to change behaviour. That preset moved no played value in 500 iterations.
Fixes two stale comments that described the wrong engine. One claimed the TT
static eval is 8 bits with +/-508cp granularity; it has been 16 full bits
since 4.0. The other gave the negative-extension defaults as 1 and 0=off
while the globals are 2 and 3. This matters more than it looks: stage 6, worth
+18 Elo, came from exactly this — a correct comment above a constant that did
something else.
Widens the time-manager cap to i64. optimum * TMMaxMult overflows int32 above
roughly two hours per move: unreachable in a game, reachable in analysis, free
to close.
Two items from the list were examined and NOT done. MB_VERIFY is already
wired — the note was itself stale. And replacing redamp with nn_finalize is
not a refactor: nn_finalize also applies EvalScale and clamps to the TB band,
so it would scale an already-scaled value twice, and even the damping alone
differs on integers because v - v*r/199 truncates at a different point than
v*(199-r)/199. If it is ever done, it needs a toggle and an SPRT.
bench unchanged at 252,074.

+++ source/uci_mt.cpp
@@ -332,10 +332,14 @@ void parse_go(char* command)
                                - (long long)overhead * (2 + mtg);
                 if (pool < 1) pool = 1;
                 optimum = (int)(pool / mtg * g_tmv2_opt_pct / 100);
-                maximum = optimum * g_tm_max_mult / 100;
+                // i64: `optimum * g_tm_max_mult` sfora int32 quando optimum supera
+                // ~2,1e9/mult. Con mult=300 succede sopra ~7,15e6 ms = ~2 ore per
+                // mossa: irraggiungibile in partita, ma raggiungibile in analisi
+                // (`go infinite` con orologi enormi) e gratis da chiudere.
+                maximum = (int)((long long)optimum * g_tm_max_mult / 100);
             } else {
                 optimum = remaining / mtg + inc * g_tm_inc_frac / 100;    // quota base + % incremento (tunable)
-                maximum = optimum * g_tm_max_mult / 100;                   // burst su posizioni difficili (tunable)
+                maximum = (int)((long long)optimum * g_tm_max_mult / 100); // burst su posizioni difficili (tunable)
             }
 
             // TM v2 Q-06 (Caissa): l'avversario ha giocato la risposta PREDETTA dalla
